Splitting pageProperties and pageProperty
