+++ webapp/src/components/centerPanelPages/pageIconSelector.tsx
@@ -27,12 +27,14 @@ const PageIconSelector = React.memo((props: Props) => {
         mutator.changeBlockIcon(page.boardId, page.id, page.fields?.icon, emoji)
         document.body.click()
     }, [page?.boardId, page?.id, page?.fields?.icon, board.id, board.icon])
+
     const onAddRandomIcon = useCallback(() => {
         if (page.parentId === '') {
             mutator.changeBoardIcon(board.id, board.icon, BlockIcons.shared.randomIcon())
         }
         mutator.changeBlockIcon(page.boardId, page.id, page.fields?.icon, BlockIcons.shared.randomIcon()), [page.boardId, page.id, page.fields?.icon]
     }, [page?.boardId, page?.id, page?.fields?.icon, board.id, board.icon])
+
     const onRemoveIcon = useCallback(() => {
         if (page.parentId === '') {
             mutator.changeBoardIcon(board.id, board.icon, '', 'remove page icon')
