+++ webapp/src/components/centerPanelPages/pageProperties.scss
@@ -1,3 +1,5 @@
+@import '../../styles/z-index';
+
 .PageProperties {
     @include z-index(center-panel);
     display: flex;
@@ -8,23 +10,4 @@
         display: none;
     }
 
-    .octo-propertyrow {
-        display: flex;
-        align-items: center;
-        border-left: 1px solid rgba(var(--center-channel-color-rgb), 0.16);
-        margin-right: 12px;
-        padding-left: 12px;
-        gap: 4px;
-        margin-bottom: 8px;
-
-        &:first-child {
-            margin-left: -10px;
-            padding-left: 0;
-            border: 0;
-        }
-
-        .Label {
-            margin: 0;
-        }
-    }
 }
